+++ src/mssql/commands.js
@@ -331,6 +331,24 @@ function commands() {
 
     getCatalog: knex => knex.raw('select db_name() as [name]'),
 
+    getVersionInfo: (knex) => {
+      const sql = `SELECT
+                    CASE 
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '8%' THEN 'SQL2000'
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '9%' THEN 'SQL2005'
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '10.0%' THEN 'SQL2008'
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '10.5%' THEN 'SQL2008 R2'
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '11%' THEN 'SQL2012'
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '12%' THEN 'SQL2014'
+                       WHEN CONVERT(VARCHAR(128), SERVERPROPERTY ('productversion')) like '13%' THEN 'SQL2016'     
+                       ELSE 'unknown'
+                    END AS majorVersion,
+                    convert(varchar(1000), SERVERPROPERTY('Edition')) AS edition,
+                    convert(varchar(1000), SERVERPROPERTY('ProductLevel')) AS productLevel,
+                    convert(varchar(1000), SERVERPROPERTY('ProductVersion')) AS productVersion`;
+      return knex.raw(sql);
+    },
+
     getSchemas: (knex) => {
       const sql = `SELECT distinct a.schema_id as id, a.name as [name] 
                   FROM sys.schemas a