+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ### [Unreleased][unreleased]
 
-- Allow to execute tests in a browser environment (PLANNED)
+- PostgreSQL schema export support
+
+### [v0.0.2] - 2016-12-11
+
+- Added MySQL schema export support
 
 ### [v0.0.1] - 2016-11-11
 
 - Scaffolding of project using [Babel Starter Kit](https://www.kriasoft.com/babel-starter-kit/)
 - Added [Travis-CI](https://travis-ci.org/) integration
+- Initial SqlServer schema export support